+++ Dockerfile
@@ -1,0 +1,18 @@
+FROM python:3.12-slim
+LABEL maintainer="hryhorii.bilous@gmail.com"
+
+ENV PYTHONBUFFERED 1
+
+WORKDIR app/
+
+COPY requirements.txt requirements.txt
+RUN pip install -r requirements.txt
+
+COPY . .
+
+RUN adduser \
+    --disabled-password \
+    --no-create-home \
+    django-user
+
+USER django-user
